Création du makefile

+++ test/utils/test_utils.py
@@ -1,4 +1,3 @@
-import numpy as np
 import pandas as pd
 from src.utils.utils import extract_categorical_indexes_from_df
 
@@ -7,7 +6,7 @@ def test_extract_categorical_indexes_from_df():
     # given
     dataframe = pd.DataFrame({'col1': [1, 2, 3], 'col2': [4, 5, 6], 'col3': ['A', 'B', 'C'], 'col4': [4, 5, 6],
                               'col5': ['A', 'B', 'C']})
-    dataframe = dataframe.astype({'col1': np.int, 'col2': np.int, 'col3': 'category', 'col4': np.int,
+    dataframe = dataframe.astype({'col1': int, 'col2': int, 'col3': 'category', 'col4': int,
                                   'col5': 'category'})
 
     # when
